Refactor typed arrays in core and worker files for improved memory efficiency
- Replaced Int32Array with Uint16Array in DecorationPool, GameObject, Scene, and various worker files to optimize memory usage.
- Updated initialization values for free list and sprite pool indices to use 0xFFFF instead of -1, enhancing consistency across the codebase.

+++ src/core/DecorationPool.js
@@ -29,7 +29,7 @@ export class DecorationPool {
     this.initialized = true;
 
     // Initialize free list with all indices (LIFO stack)
-    this.freeList = new Int32Array(maxDecorations);
+    this.freeList = new Uint16Array(maxDecorations);
     for (let i = 0; i < maxDecorations; i++) {
       this.freeList[i] = i;
     }
